pyproject.toml
[tool.poetry]
name = "new-project"
version = "0.1.0"
description = ""
authors = ["Lunteg <ozerov-danil@yandex.ru>"]
readme = "README.md"
packages = [{include = "new_project"}]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.3"
python-dotenv = "^0.21.1"
psycopg2 = "^2.9.5"
pandas = "^1.5.3"
sqlalchemy = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
